Add data storage via SQLite (Room)
Added the use of a database to store the history of
queries and their results. SQLite database management is
implemented through the ROOM library to reduce the
amount of code and simplify the database management
process.

The RequestResult class is created to describe the table
and the data that will be stored in it
Created the RequestResultDao interface, which describes
methods and ways to interact with the database
Created the abstract class FactsAboutNumber_Database to
create and access the database.

In the MainActivity class, the logic of displaying the
received facts is displayed in a separate method
showResult(String result). It is used when loading data
from the database when creating an Activity and when
receiving the result of a query through the API